Add a container to wrap 1st container's content
This is done to align the content (title, body and button)
in the desired way. The result is the main container (for everything
in that section), a "left container" that expands to fill the middle
portion horizontally, a content container inside of the previous one
(contains the title, body and button) which can be aligned horizontally
in the desired manner.