Modify some 1st container's rules
Set the left content div's justify-content attribute to flex-end so
that it is closer to the image place-holder. Add a right margin the
the left content's content so that it is a bit separated from the
image placeholder. This is done so that it is more accurate to the
desired outcome.